Se activó usuario nuevamente + se lista tarjetas por div

+++ BancaTodoFrontend/src/app/cliente/components/lista-cliente/lista-cliente.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ToastrService } from 'ngx-toastr';
 import { Cliente } from 'src/app/cliente/models/cliente';
 import { ClienteService } from 'src/app/cliente/services/cliente.service';
+import Swal from 'sweetalert2';
 
 
 @Component({
@@ -19,7 +20,7 @@ export class ListaClienteComponent implements OnInit {
   ngOnInit(): void {
     this.cargarClientes();
 
-    console.log(localStorage.getItem('tutorial'));
+    //console.log(localStorage.getItem('tutorial'));
   }
 
   cargarClientes(): void{
@@ -28,37 +29,62 @@ export class ListaClienteComponent implements OnInit {
       respuesta => {
         if (respuesta.peticionExitosa){
           this.clientes = respuesta.datos;
-          localStorage.setItem('tutorial', 'Como utilizar el LocalStorage en Angular');
+          //localStorage.setItem('tutorial', 'Como utilizar el LocalStorage en Angular');
         }
 
-      }, err => {console.log(err.error.mensaje)}
+      }, err => {
+        Swal.fire(
+          'Error!',
+          err.error.mensaje,
+          'error'
+        );
+      }
     )
   }
 
   borrar(id: number){
-    this.clienteService.delete(id).subscribe(
-      respuesta=>{
-        if (respuesta.peticionExitosa) {
-          console.log(respuesta.mensaje.charAt(0));
-          if (respuesta.mensaje.charAt(0) == "0") {
-            this.toastr.success(respuesta.mensaje, '¡Satisfactoria!', {
-              timeOut: 4000,
-              positionClass: 'toast-top-center',
-            });
-            this.cargarClientes();
-          }else{
-            this.toastr.warning(respuesta.mensaje, '¡Info!', {
-              timeOut: 4000,
-              positionClass: 'toast-top-center',
-            });
+
+
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.clienteService.delete(id).subscribe(
+          respuesta=>{
+            if (respuesta.peticionExitosa) {
+
+              if (respuesta.mensaje.charAt(0) == "0") {
+                Swal.fire(
+                  'Deleted!',
+                  'User has been deleted.',
+                  'success'
+                );
+                this.cargarClientes();
+              }else{
+                Swal.fire({
+                  icon: 'error',
+                  title: '¡Info!',
+                  text: respuesta.mensaje,
+                })
+              }
+            }
+          },
+          err =>{
+            Swal.fire({
+              icon: 'error',
+              title: '¡Error!',
+              text: err.error.mensaje,
+            })
+
           }
-        }
-      },
-      err =>{
-        this.toastr.error(err.error.mensaje,'Fail',{
-          timeOut:4000, positionClass: 'toast-top-center',
-        });
+        )
       }
-    )
+    });
   }
 }
